Move form task to a command
