+++ src/Command/FormCommand.php
@@ -13,14 +13,12 @@ declare(strict_types=1);
  * @since         1.0.9
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Bake\Shell\Task;
+namespace Bake\Command;
 
 /**
- * ShellHelper code generator.
- *
- * @property \Bake\Shell\Task\TestTask $Test
+ * Form code generator.
  */
-class FormTask extends SimpleBakeTask
+class FormCommand extends SimpleBakeCommand
 {
     /**
      * Task name used in path generation.
