+++ src/Plugin.php
@@ -19,6 +19,7 @@ use Bake\Command\BehaviorCommand;
 use Bake\Command\CellCommand;
 use Bake\Command\CommandCommand;
 use Bake\Command\ComponentCommand;
+use Bake\Command\FormCommand;
 use Bake\Command\TestCommand;
 use Bake\Shell\BakeShell;
 use Cake\Console\CommandCollection;
@@ -71,6 +72,7 @@ class Plugin extends BasePlugin
         $commands->add('bake cell', CellCommand::class);
         $commands->add('bake command', CommandCommand::class);
         $commands->add('bake component', ComponentCommand::class);
+        $commands->add('bake form', FormCommand::class);
         $commands->add('bake test', TestCommand::class);
 
         // Add shell for incomplete tasks and backwards compatibility discover.
