+++ tests/TestCase/Shell/BakeShellTest.php
@@ -208,15 +208,8 @@ class BakeShellTest extends TestCase
             '- controller',
             '- custom_controller',
             '- fixture',
-            '- form',
-            '- helper',
-            '- mailer',
-            '- middleware',
             '- model',
             '- plugin',
-            '- shell',
-            '- shell_helper',
-            '- task',
             '- template',
             '- test',
             '- twig_template',
@@ -260,15 +253,8 @@ class BakeShellTest extends TestCase
         $this->Shell->loadTasks();
         $expected = [
             'Bake.Fixture',
-            'Bake.Form',
-            'Bake.Helper',
-            'Bake.Mailer',
-            'Bake.Middleware',
             'Bake.Model',
             'Bake.Plugin',
-            'Bake.Shell',
-            'Bake.ShellHelper',
-            'Bake.Task',
             'Bake.Test',
             'Bake.Template',
             'Controller',
